+++ gazon-service/src/main/java/com/gitlab/controller/ExampleRestController.java
@@ -2,7 +2,6 @@ package com.gitlab.controller;
 
 import com.gitlab.controller.api.ExampleRestApi;
 import com.gitlab.dto.ExampleDto;
-import com.gitlab.mapper.ExampleMapper;
 import com.gitlab.model.Example;
 import com.gitlab.service.ExampleService;
 import lombok.RequiredArgsConstructor;
@@ -24,8 +23,6 @@ public class ExampleRestController implements ExampleRestApi {
 
     private final ExampleService exampleService;
 
-    private final ExampleMapper exampleMapper;
-
     @Override
     public ResponseEntity<Page<ExampleDto>> getPage(Integer page, Integer size) {
         if (page == null || size == null) {
@@ -44,42 +41,37 @@ public class ExampleRestController implements ExampleRestApi {
     }
 
     private ResponseEntity<Page<ExampleDto>> createUnPagedResponse() {
-        var examples = exampleService.findAll();
-        if (examples.isEmpty()) {
+        var exampleDtos = exampleService.findAllDto();
+        if (exampleDtos.isEmpty()) {
             return ResponseEntity.noContent().build();
         }
-        return ResponseEntity.ok(new PageImpl<>(examples.stream().map(exampleMapper::toDto).toList()));
+        return ResponseEntity.ok(new PageImpl<>(exampleDtos));
     }
 
     private ResponseEntity<Page<ExampleDto>> createPagedResponse(Page<Example> examplePage) {
-        var exampleDtoPage = new PageImpl<>(
-                examplePage.getContent().stream().map(exampleMapper::toDto).toList(),
-                examplePage.getPageable(),
-                examplePage.getTotalElements()
-        );
+        var exampleDtoPage = exampleService.getPageDto(examplePage.getPageable().getPageNumber(), examplePage.getPageable().getPageSize());
         return ResponseEntity.ok(exampleDtoPage);
     }
 
     @Override
     public ResponseEntity<ExampleDto> get(Long id) {
-        return exampleService.findById(id)
-                .map(value -> ResponseEntity.ok(exampleMapper.toDto(value)))
+        return exampleService.findByIdDto(id)
+                .map(value -> ResponseEntity.ok(value))
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @Override
     public ResponseEntity<ExampleDto> create(ExampleDto exampleDto) {
+        ExampleDto savedExampleDto = exampleService.saveDto(exampleDto);
         return ResponseEntity.status(HttpStatus.CREATED)
-                .body(exampleMapper
-                        .toDto(exampleService
-                                .save(exampleMapper
-                                        .toEntity(exampleDto))));
+                .body(savedExampleDto);
     }
 
     @Override
     public ResponseEntity<ExampleDto> update(Long id, ExampleDto exampleDto) {
-        return exampleService.update(id, exampleMapper.toEntity(exampleDto))
-                .map(example -> ResponseEntity.ok(exampleMapper.toDto(example)))
+        Optional<ExampleDto> updatedExampleDto = exampleService.updateDto(id, exampleDto);
+        return updatedExampleDto
+                .map(dto -> ResponseEntity.ok(dto))
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }